Tests added: testParameterizedBuildWithChoices
Methods added:
configureParameterizedBuild, addTimeStampToConsoleOutput, createUniqueTextValue, testMoveFreestyleToFolder

Classes deleted:
FreestyleProject23Test, FreestyleProject24Test, FreestyleProject27Test, FreestyleProject28Test, FreestyleProject32Test, FreestyleProject88Test, FreestyleOmegakotTest, FreestyleProjectPopovTest, FreestyleProjectIldanaTest, FreestyleProjectThTest, FreestyleMoveToFolderTest
